feat: move Replay tour to a dev-only Developer settings section
Replay tour now lives in a dedicated Developer settings section gated on
import.meta.env.DEV, so it appears under `bun tauri dev` but not in
production builds. Removed the "Getting started" group from General.
